Refactored My Mapping Implementation
This implementation contains the occupancy grid, with the methods
incOccGrid(int x, int y), decOccGrid(int x, int y), incCountGraph(int x,
int y) and calcCellProb(int x, int y).

getMap() and printMap() now take three parameters - int[][], int, int.
This is to re-use code when we print out the Occupancy Grid or the Main
map that we had.

the scanX() methods use xTurn() to define the orientation of the cell
modified in updateMap() - think of it as identifying where you are
looking and you then check if the cell where you're looking is empty or
not.

updateMap() now takes two int variables - distance and direction.